remove with lock versions
