+++ operator-framework/src/test/java/io/javaoperatorsdk/operator/baseapi/statuscache/StatusPatchPrimaryCacheIT.java
@@ -1,4 +1,4 @@
-package io.javaoperatorsdk.operator.baseapi.statuscache.withlock;
+package io.javaoperatorsdk.operator.baseapi.statuscache;
 
 import java.time.Duration;
 
@@ -11,19 +11,19 @@ import io.javaoperatorsdk.operator.junit.LocallyRunOperatorExtension;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.awaitility.Awaitility.await;
 
-public class StatusPatchCacheWithLockIT {
+class StatusPatchPrimaryCacheIT {
 
   public static final String TEST_1 = "test1";
 
   @RegisterExtension
   LocallyRunOperatorExtension extension =
       LocallyRunOperatorExtension.builder()
-          .withReconciler(StatusPatchCacheWithLockReconciler.class)
+          .withReconciler(StatusPatchPrimaryCacheReconciler.class)
           .build();
 
   @Test
   void testStatusAlwaysUpToDate() {
-    var reconciler = extension.getReconcilerOfType(StatusPatchCacheWithLockReconciler.class);
+    var reconciler = extension.getReconcilerOfType(StatusPatchPrimaryCacheReconciler.class);
 
     extension.create(testResource());
 
@@ -39,10 +39,10 @@ public class StatusPatchCacheWithLockIT {
             });
   }
 
-  StatusPatchCacheWithLockCustomResource testResource() {
-    var res = new StatusPatchCacheWithLockCustomResource();
+  StatusPatchPrimaryCacheCustomResource testResource() {
+    var res = new StatusPatchPrimaryCacheCustomResource();
     res.setMetadata(new ObjectMetaBuilder().withName(TEST_1).build());
-    res.setSpec(new StatusPatchCacheWithLockSpec());
+    res.setSpec(new StatusPatchPrimaryCacheSpec());
     return res;
   }
 }
