+++ operator-framework/src/test/java/io/javaoperatorsdk/operator/baseapi/statuscache/StatusPatchPrimaryCacheSpec.java
@@ -1,4 +1,4 @@
-package io.javaoperatorsdk.operator.baseapi.statuscache.primarycache;
+package io.javaoperatorsdk.operator.baseapi.statuscache;
 
 public class StatusPatchPrimaryCacheSpec {
