--- operator-framework/src/test/java/io/javaoperatorsdk/operator/baseapi/statuscache/withlock/StatusPatchCacheWithLockSpec.java
@@ -1,14 +1,0 @@
-package io.javaoperatorsdk.operator.baseapi.statuscache.withlock;
-
-public class StatusPatchCacheWithLockSpec {
-
-  private int counter = 0;
-
-  public int getCounter() {
-    return counter;
-  }
-
-  public void setCounter(int counter) {
-    this.counter = counter;
-  }
-}
