--- operator-framework/src/test/java/io/javaoperatorsdk/operator/baseapi/statuscache/withlock/StatusPatchCacheWithLockStatus.java
@@ -1,15 +1,0 @@
-package io.javaoperatorsdk.operator.baseapi.statuscache.withlock;
-
-public class StatusPatchCacheWithLockStatus {
-
-  private Integer value = 0;
-
-  public Integer getValue() {
-    return value;
-  }
-
-  public StatusPatchCacheWithLockStatus setValue(Integer value) {
-    this.value = value;
-    return this;
-  }
-}
